Make Restaurant Properties public Added nearby restaurant locator

+++ fillbelly_test.go
@@ -14,15 +14,39 @@ func isEnvTest() bool{
 
 func ResetDB() {
 	if(isEnvTest()) {
-		clearRestaurantTable()
-		// createRestaurantTable()
-		// createCategoryTable()
+		dropRestaurantTable()
+		dropCategoryTable()
+		createRestaurantTable()
+		createCategoryTable()
 	} else {
 		fmt.Printf("ENV is not test, aborting db reset");
 	}
 }
 
+func dropRestaurantTable(){
+	if(!isEnvTest()) { return }
+
+	db, _ := sql.Open("postgres", DbConnectionString());
+	quoted := pq.QuoteIdentifier("restaurants")
+	_, err := db.Exec(fmt.Sprintf("DROP TABLE %s", quoted))
+	if err != nil {
+		fmt.Printf("Error: ", err)
+	}
+}
+
+func dropCategoryTable(){
+	if(!isEnvTest()) { return }
+
+	db, _ := sql.Open("postgres", DbConnectionString());
+	quoted := pq.QuoteIdentifier("categories")
+	_, err := db.Exec(fmt.Sprintf("DROP TABLE %s", quoted))
+	if err != nil {
+		fmt.Printf("Error: ", err)
+	}
+}
+
 func clearRestaurantTable(){
+	if(!isEnvTest()) { return }
 	db, _ := sql.Open("postgres", DbConnectionString());
 	quoted := pq.QuoteIdentifier("restaurants")
 	_, err := db.Exec(fmt.Sprintf("DELETE FROM %s", quoted))
@@ -33,7 +57,7 @@ func clearRestaurantTable(){
 
 func createRestaurantTable(){
 	db, _ := sql.Open("postgres", DbConnectionString());
-	_, err := db.Exec(fmt.Sprintf("CREATE TABLE restaurants ( id serial primary key, name varchar(200), rating integer, longitude varchar(50), latitude varchar(50), id_category int);"))
+	_, err := db.Exec(fmt.Sprintf("CREATE TABLE restaurants ( id serial primary key, name varchar(200), rating integer, longitude float, latitude float, id_category int);"))
 	if err != nil {
 		fmt.Printf("Error: ", err)
 	}
@@ -47,35 +71,96 @@ func createCategoryTable(){
 	}
 }
 
+func TestRestaurantLocator(t *testing.T) {
+	if(!isEnvTest()) { return }
+	ResetDB()
+
+	categoryName := "Seafood"
+	idCategory := InsertNewCategory(Category{ name: categoryName })
+	
+	restaurantName := "Puvlic"
+	restaurantRating := 9
+	restaurantLatitude := -6.115894
+	restaurantLongitude := 106.7867153
+	restaurantCategory := idCategory
+
+	InsertNewRestaurant(Restaurant{
+		Name: restaurantName,
+		Rating: restaurantRating,
+		Longitude: restaurantLongitude,
+		Latitude: restaurantLatitude,
+		Id_category: restaurantCategory,
+	})
+
+	farRestaurantName := "Masda"
+	farRestaurantRating := 9
+	farRestaurantLatitude := -6.174722
+	farRestaurantLongitude := 106.7903383
+	farRestaurantCategory := idCategory
+
+	InsertNewRestaurant(Restaurant{
+		Name: farRestaurantName,
+		Rating: farRestaurantRating,
+		Longitude: farRestaurantLongitude,
+		Latitude: farRestaurantLatitude,
+		Id_category: farRestaurantCategory,
+	})
+
+	latitude := -6.115734
+	longitude := 106.7916073
+
+	results := getNearbyRestaurants(latitude, longitude)
+	var matched bool
+	matched = false
+	for _, element := range results {
+		if(element.Name == farRestaurantName) {
+			t.Errorf("Restaurant %s should not be included.", farRestaurantName);
+		}
+	}
+
+	for _, element := range results {
+		if(element.Name == restaurantName) {
+			matched = true
+		}
+
+	}
+	if !matched {
+		t.Errorf("Restaurant %s should be included.", restaurantName);
+	}
+}
+
 func TestRestaurantModel(t *testing.T) {
 	if(!isEnvTest()) { return }
 	ResetDB()
 
-	idCategory := InsertNewCategory(Category{ name: "Seafood" })
+	categoryName := "Seafood"
+
+	idCategory := InsertNewCategory(Category{ name: categoryName })
 
 	restaurantName := "Puvlic"
 	restaurantRating := 9
-	restaurantLongitude := "6.998"
-	restaurantLatitude := "688.998"
+	restaurantLongitude := 6.998
+	restaurantLatitude := 688.998
 	restaurantCategory := idCategory
 
 	InsertNewRestaurant(Restaurant{
-		name: restaurantName,
-		rating: restaurantRating,
-		longitude: restaurantLongitude,
-		latitude: restaurantLatitude,
-		id_category: restaurantCategory,
+		Name: restaurantName,
+		Rating: restaurantRating,
+		Longitude: restaurantLongitude,
+		Latitude: restaurantLatitude,
+		Id_category: restaurantCategory,
 	})
 
 	all := GetAllRestaurant()
 
 	first := all[0]
 
-	if( first.name != restaurantName ||
-		first.id_category != restaurantCategory ||
-		first.rating != restaurantRating ||
-		first.longitude != restaurantLongitude ||
-		first.latitude != restaurantLatitude) {
+	if( first.Name != restaurantName ||
+		first.Id_category != restaurantCategory ||
+		first.Rating != restaurantRating ||
+		first.Longitude != restaurantLongitude ||
+		first.Category_name != categoryName ||
+		first.Latitude != restaurantLatitude) {
 		t.Errorf("Some element did not match")
 	}
 }
